requirements.txt
numpy
librosa
pysoundfile
matplotlib
scikit-learn
tensorflow
keras
